separate out the test into traversal_test.go
